# nexus-hub-clone/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Nexus OS API - Knowledge Management + AI Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /memory:
    get:
      operationId: listMemories
      tags: [memory]
      summary: List AI memories
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [short_term, long_term, skill]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Memories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Memory"
    post:
      operationId: createMemory
      tags: [memory]
      summary: Create a memory entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMemoryBody"
      responses:
        "201":
          description: Created memory
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Memory"
  /memory/{memoryId}:
    delete:
      operationId: deleteMemory
      tags: [memory]
      summary: Delete a memory
      parameters:
        - name: memoryId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /memory/search:
    post:
      operationId: searchMemory
      tags: [memory]
      summary: Semantic search through memories
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemorySearchBody"
      responses:
        "200":
          description: Matching memories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Memory"

  # ── SKILLS ───────────────────────────────────────────────────────────────────
  /skills:
    get:
      operationId: listSkills
      tags: [skills]
      summary: List AI skills
      parameters:
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Skills
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Skill"
    post:
      operationId: createSkill
      tags: [skills]
      summary: Create a custom skill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSkillBody"
      responses:
        "201":
          description: Created skill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Skill"
  /skills/{skillId}/run:
    post:
      operationId: runSkill
      tags: [skills]
      summary: Execute a skill
      parameters:
        - name: skillId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunSkillBody"
      responses:
        "200":
          description: Skill result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillRunResult"

  # ── DATABASES ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Memory:
      type: object
      properties:
        id:
          type: string
        content:
          type: string
        type:
          type: string
          enum: [short_term, long_term, skill, preference]
        source:
          type: string
        importance:
          type: integer
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [id, content, type]
    CreateMemoryBody:
      type: object
      properties:
        content:
          type: string
        type:
          type: string
          enum: [short_term, long_term, skill, preference]
        source:
          type: string
        importance:
          type: integer
      required: [content, type]
    MemorySearchBody:
      type: object
      properties:
        query:
          type: string
        limit:
          type: integer
          default: 10
      required: [query]

    # ── SKILL ─────────────────────────────────────────────────────────────────
    Skill:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        category:
          type: string
        prompt:
          type: string
        isBuiltIn:
          type: boolean
        icon:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, name, description, category, isBuiltIn]
    CreateSkillBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
        prompt:
          type: string
        icon:
          type: string
      required: [name, description, category, prompt]
    RunSkillBody:
      type: object
      properties:
        input:
          type: string
        pageId:
          type: string
        providerId:
          type: string
      required: [input]
    SkillRunResult:
      type: object
      properties:
        output:
          type: string
        skillId:
          type: string
        tokensUsed:
          type: integer
        executionMs:
          type: integer
      required: [output, skillId]

    # ── DATABASE ──────────────────────────────────────────────────────────────
...